+++ R/rebuild_cached_data.R
@@ -21,6 +21,7 @@
 #' cadrs_new <- radrs(cadsl, seed = 1, na_vars = list()) %>% `attr<-`("creation date", Sys.Date())
 #' cadtte_new <- radtte(cadsl, seed = 1, na_vars = list()) %>% `attr<-`("creation date", Sys.Date())
 #' cadvs_new <- radvs(cadsl, seed = 1, na_vars = list()) %>% `attr<-`("creation date", Sys.Date())
+#' cadeg_new <- radeg(cadsl, seed = 1, na_vars = list()) %>% `attr<-`("creation date", Sys.Date())
 #'
 #' # use diffdf package: diffdf() to compare cached vs new and review/verify differences
 #' \dontrun{
@@ -38,6 +39,7 @@
 #' cadrs <- cadrs_new
 #' cadtte <- cadtte_new
 #' cadvs <- cadvs_new
+#' cadeg <- cadeg_new
 #'
 #' # once diff review is completed then update cache
 #' save(cadsl, file = "data/cadsl.RData", compress = "xz")
@@ -51,3 +53,4 @@
 #' save(cadrs, file = "data/cadrs.RData", compress = "xz")
 #' save(cadtte, file = "data/cadtte.RData", compress = "xz")
 #' save(cadvs, file = "data/cadvs.RData", compress = "xz")
+#' save(cadeg, file = "data/cadeg.RData", compress = "xz")